+++ NEWS.md
@@ -11,11 +11,14 @@
 
 ### Bug fixes
 * Add retry for search operations ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+* Skip background processing on BadSqlGrammarException ([MSEARCH-1214](https://folio-org.atlassian.net/browse/MSEARCH-1214))
 
 ### Tech Dept
 * Add `@TestRailCase` annotation for linking integration tests to TestRail cases
 * Add integration test coverage for SUDOC call-number browse type filtering (TestRail C627509)
 * Add integration test coverage for LC call-number browse with empty config returning all types (TestRail C627500)
+* Add integration test coverage for NLM call-number browse with empty config returning all types (TestRail C627501)
+* Add integration test coverage for Other scheme call-number browse with empty config returning all types (TestRail C627502)
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`